clock job format changes
